The-Graph-Data-Injest/requirements.txt
functions-framework==3.*
requests==2.30.0
google-cloud-bigquery==3.10.0
